# frontend/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Liveness probe",
        "tags": [
          "Health"
        ]
      }
    },
    "/api/users": {
      "post": {
        "operationId": "UserController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "550e8400-e29b-41d4-a716-446655440000"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "https://my-api-doc.dev/problems/validation-error"
                    },
                    "title": {
                      "type": "string",
                      "example": "Validation Error"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    },
                    "detail": {
                      "type": "string",
                      "example": "One or more fields failed validation."
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string",
                            "example": "email"
                          },
                          "message": {
                            "type": "string",
                            "example": "email must be an email"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "about:blank"
                    },
                    "title": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "number",
                      "example": 401
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Not a System Admin",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "title": {
                      "type": "string",
                      "example": "Forbidden"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "https://my-api-doc.dev/problems/username-already-exists"
                    },
                    "title": {
                      "type": "string",
                      "example": "Username Already Exists"
                    },
                    "status": {
                      "type": "number",
                      "example": 409
                    },
                    "detail": {
                      "type": "string",
                      "example": "A user already exists with username sina.q"
                    },
                    "username": {
                      "type": "string",
                      "example": "sina.q"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a new user",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UserController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                      },
                      "name": {
                        "type": "string",
                        "example": "Sina Ghadrdan"
                      },
                      "username": {
                        "type": "string",
                        "example": "sina.q"
                      },
                      "role": {
                        "type": "string",
                        "enum": [
                          "system_admin",
                          "management",
                          "qc_inspector",
                          "reporter"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "about:blank"
                    },
                    "title": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "number",
                      "example": 401
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Not a System Admin",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "title": {
                      "type": "string",
                      "example": "Forbidden"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List every registered user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "User updated successfully"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "https://my-api-doc.dev/problems/validation-error"
                    },
                    "title": {
                      "type": "string",
                      "example": "Validation Error"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    },
                    "detail": {
                      "type": "string",
                      "example": "One or more fields failed validation."
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string",
                            "example": "email"
                          },
                          "message": {
                            "type": "string",
                            "example": "email must be an email"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "about:blank"
                    },
                    "title": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "number",
                      "example": 401
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Not a System Admin",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "title": {
                      "type": "string",
                      "example": "Forbidden"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "https://my-api-doc.dev/problems/entity-not-found"
                    },
                    "title": {
                      "type": "string",
                      "example": "Entity not found"
                    },
                    "status": {
                      "type": "number",
                      "example": 404
                    },
                    "detail": {
                      "type": "string",
                      "example": "Entity not found with id 550e8400-e29b-41d4-a716-446655440000"
                    },
                    "entityId": {
                      "type": "string",
                      "example": "550e8400-e29b-41d4-a716-446655440000"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "https://my-api-doc.dev/problems/cannot-change-own-role"
                    },
                    "title": {
                      "type": "string",
                      "example": "Cannot Change Own Role"
                    },
                    "status": {
                      "type": "number",
                      "example": 409
                    },
                    "detail": {
                      "type": "string",
                      "example": "A System Admin cannot change their own role"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit an existing user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UserController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "about:blank"
                    },
                    "title": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "number",
                      "example": 401
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Not a System Admin",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "title": {
                      "type": "string",
                      "example": "Forbidden"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "https://my-api-doc.dev/problems/entity-not-found"
                    },
                    "title": {
                      "type": "string",
                      "example": "Entity not found"
                    },
                    "status": {
                      "type": "number",
                      "example": 404
                    },
                    "detail": {
                      "type": "string",
                      "example": "Entity not found with id 550e8400-e29b-41d4-a716-446655440000"
                    },
                    "entityId": {
                      "type": "string",
                      "example": "550e8400-e29b-41d4-a716-446655440000"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "https://my-api-doc.dev/problems/validation-error"
                    },
                    "title": {
                      "type": "string",
                      "example": "Validation Error"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    },
                    "detail": {
                      "type": "string",
                      "example": "One or more fields failed validation."
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string",
                            "example": "email"
                          },
                          "message": {
                            "type": "string",
                            "example": "email must be an email"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "https://my-api-doc.dev/problems/invalid-credentials"
                    },
                    "title": {
                      "type": "string",
                      "example": "Invalid Credentials"
                    },
                    "status": {
                      "type": "number",
                      "example": 401
                    },
                    "detail": {
                      "type": "string",
                      "example": "Invalid username or password provided."
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Log in with a username and password",
        "tags": [
          "Auth"
        ]
      }
    }
  },
  "info": {
    "title": "API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Role": {
        "type": "string",
        "enum": [
          "system_admin",
          "management",
          "qc_inspector",
          "reporter"
        ]
      },
      "RegisterUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sina Ghadrdan"
          },
          "username": {
            "type": "string",
            "example": "sina.q"
          },
          "password": {
            "type": "string",
            "example": "Passw0rd!"
          },
          "role": {
            "example": "qc_inspector",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          }
        },
        "required": [
          "name",
          "username",
          "password",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sina Ghadrdan"
          },
          "username": {
            "type": "string",
            "example": "sina.q"
          },
          "password": {
            "type": "string",
            "example": "NewPassw0rd!"
          },
          "role": {
            "example": "management",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          }
        }
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "sina.q"
          },
          "password": {
            "type": "string",
            "example": "Passw0rd!"
          }
        },
        "required": [
          "username",
          "password"
        ]
      }
    }
  }
}